+++ test/convlayers-tests.jl
@@ -9,9 +9,7 @@
 
 @test size(tm.UpConvK2(3, 1)(x3)) == (512,512,1,1)
 
-@test size(tm.ConvTrK2(3, 1)(x3)) == (256,256,1,1)
-@test size(tm.ConvTrK2(3, 1; stride=2)(x3)) == (512,512,1,1)
-
+@test size(tm.ConvTrK2(3, 1)(x3)) == (512,512,1,1)
 @test size(tm.ConvTrK4(3, 1)(x3)) == (512,512,1,1)
 
 @test size(tm.DilatedConvK3(3,1; stride=1, dilation=1)(x3)) == (256,256,1,1)